commaData: the central data object for the comma package

commaData is an S4 class that extends SummarizedExperiment to store genome-wide bacterial methylation data from Oxford Nanopore sequencing. It is the central object accepted and returned by all comma analysis functions.

Value

An object of class commaData. Use commaData to construct instances.

Details

The class stores methylation data in two assay matrices (accessible via assay):

"methylation"

Beta values (proportion of reads called methylated, range 0–1). Sites with coverage below the min_coverage threshold are stored as NA.

"coverage"

Integer read depth at each site.

Per-site metadata is in rowData(object) and includes at minimum: chrom, position, strand, mod_type.

Per-sample metadata is in colData(object) and includes at minimum: sample_name, condition, replicate.

Slots

genomeInfo

Named integer vector of chromosome sizes c(chromosome name = length in bp).

annotation

GRanges of genomic features loaded from a GFF3 or BED file. May be an empty GRanges if no annotation was provided.

motifSites

GRanges of all instances of the user-specified sequence motif in the genome (e.g., all GATC sites). May be an empty GRanges if no motif was specified.

See also